+++ packages/data/js/variables/linear-gradient.js
@@ -4,6 +4,7 @@
  */
 import { default as memoize } from 'fast-memoize';
 import { select } from '@wordpress/data';
+import { __ } from '@wordpress/i18n';
 
 /**
  * Blockera dependencies
@@ -78,6 +79,21 @@ export const getLinearGradients: () => Array<VariableItem> = memoize(
 			});
 	}
 );
+
+export const getLinearGradientsTitle: () => string = memoize(function () {
+	if (isBlockTheme()) {
+		if (
+			!isUndefined(
+				getBlockEditorSettings()?.__experimentalFeatures?.color
+					?.gradients?.theme
+			)
+		) {
+			return __('Theme Linear Gradients', 'blockera');
+		}
+	}
+
+	return __('Editor Linear Gradients', 'blockera');
+});
 
 export const getLinearGradient: (id: string) => ?VariableItem = memoize(
 	function (id: string): ?VariableItem {