+++ packages/data/js/variables/spacing.js
@@ -5,6 +5,7 @@
  */
 import { default as memoize } from 'fast-memoize';
 import { select } from '@wordpress/data';
+import { __ } from '@wordpress/i18n';
 
 /**
  * Blockera dependencies
@@ -72,6 +73,21 @@ export const getSpacings: () => Array<VariableItem> = memoize(
 		});
 	}
 );
+
+export const getSpacingsTitle: () => string = memoize(function (): string {
+	if (isBlockTheme()) {
+		if (
+			!isUndefined(
+				getBlockEditorSettings()?.__experimentalFeatures?.spacing
+					?.spacingSizes?.theme
+			)
+		) {
+			return __('Theme Spacing Sizes', 'blockera');
+		}
+	}
+
+	return __('Editor Spacing Sizes', 'blockera');
+});
 
 export const getSpacing: (id: string) => ?VariableItem = memoize(function (
 	id: string