+++ packages/controls/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Unreleased
 
-## Bug Fixes
+### Bug Fixes
 - Fixed an issue where variable colors appeared lighter in Blockera Variables than in the original variables  [[🔗 Bug](https://community.blockera.ai/bugs-mdhyb8nc/post/showing-theme-colors-as-they-are-4VjzJz6qKKL9HdQ)]
+
+### Improvements
+- Improved variable picker to display accurate title indicating whether the variable is sourced from the block theme or block editor.
 
 
 ## 1.0.0 (2024-12-08)